final touches

+++ public/js/index.js
@@ -1,6 +1,7 @@
 const allCommentsContainer = document.getElementsByClassName('all-comments')[0];
+const error = document.getElementsByClassName('error')[0];
 
-const comentarios = [];
+let comentarios = [];
 
 function init() {
   getAllComentarios();
@@ -10,17 +11,31 @@ function init() {
   SearchByAuthorFormInit();
 }
 
+function displayError(message) {
+  error.style.display = 'inline-block';
+  error.innerHTML = message;
+}
+
+function hideError() {
+  error.style.display = 'none';
+  error.innerHTML = '';
+}
+
 const Comentario = ({ titulo, autor, contenido, fecha, id }) => `
-  <li>${autor} - ${titulo} - ${contenido} - ${fecha}</li>
+  <li>${autor} - ${titulo} - ${contenido} - ${fecha} - ${id}</li>
 `;
 
 async function getAllComentarios() {
   const data = await fetch('/blog-api/comentarios');
   const dataJson = await data.json();
   allCommentsContainer.innerHTML = '';
+  comentarios = [...dataJson];
   for (let comment of dataJson) {
     allCommentsContainer.innerHTML += Comentario({ ...comment });
-    comentarios.push(comment);
+    //comentarios.push(comment);
+  }
+  if (dataJson.length == 0) {
+    allCommentsContainer.innerHTML = 'No se encontraronn comentarios';
   }
 }
 
@@ -36,18 +51,27 @@ function newCommentFormInit() {
     for (let elem of inputs) {
       data[elem.name] = elem.value;
     }
+    if (!(data.contenido && data.autor && data.titulo)) {
+      displayError(
+        'Faltan parametros para crear comentario, favor de ingresar todos los parametros'
+      );
+      return;
+    } else {
+      hideError();
+      const newData = await fetch('/blog-api/nuevo-comentario', {
+        method: 'POST',
+        headers: {
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(data),
+      });
 
-    const newData = await fetch('/blog-api/nuevo-comentario', {
-      method: 'POST',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-    const newDataJSON = await newData.json();
-    allCommentsContainer.innerHTML += Comentario({ ...newDataJSON });
-    comentarios.push(newDataJSON);
+      const newDataJSON = await newData.json();
+
+      comentarios.push(newDataJSON);
+      await getAllComentarios();
+    }
   });
 }
 
@@ -64,8 +88,16 @@ function updateCommentFormInit() {
       data[elem.name] = elem.value;
     }
 
-    console.log(data);
+    if (!(data.autor || data.titulo || data.contenido)) {
+      displayError('No se ingresaron parametros, favor de ingresar al menos 1');
+      return;
+    }
 
+    if (!data.id) {
+      displayError('No se ingreso el id del comentario a cambiar');
+      return;
+    }
+    hideError();
     const newData = await fetch('/blog-api/actualizar-comentario/' + data['id'], {
       method: 'PUT',
       headers: {
@@ -76,11 +108,10 @@ function updateCommentFormInit() {
     });
     const newDataJSON = await newData.json();
 
-    const index = comentarios.find(val => val.id === data['id']);
-    comentarios[index] = newDataJSON;
-    allCommentsContainer.innerHTML = '';
-    for (let elem of comentarios) {
-      allCommentsContainer.innerHTML += Comentario({ ...elem });
+    if (newDataJSON) {
+      getAllComentarios();
+    } else {
+      displayError(`Comentario con id ${data.id} no encontrado`);
     }
   });
 }
@@ -98,16 +129,20 @@ function RemoveCommentFormInit() {
       data[elem.name] = elem.value;
     }
 
-    const newData = await fetch('/blog-api/remover-comentario/' + data['id'], {
+    if (!data.id) {
+      displayError('Favor de proveer el id del comentario a eliminar');
+      return;
+    }
+
+    const deleted = await fetch('/blog-api/remover-comentario/' + data['id'], {
       method: 'DELETE',
     });
-
-    const index = comentarios.find(val => val.id === data['id']);
-    comentarios.splice(index, 1);
-    allCommentsContainer.innerHTML = '';
-    for (let elem of comentarios) {
-      allCommentsContainer.innerHTML += Comentario({ ...elem });
+    if (deleted.status === 404) {
+      displayError('No se encontro un comentario con el id proveido');
+      return;
     }
+    hideError();
+    getAllComentarios();
   });
 }
 
@@ -125,13 +160,17 @@ function SearchByAuthorFormInit() {
       data[elem.name] = elem.value;
     }
 
+    if (!data.name) {
+      displayError('Favor de proveer el nombre del autor');
+      return;
+    }
+
     const newData = await fetch('/blog-api/comentarios-por-autor?autor=' + data['autor']);
-    if (newData.status > 210) {
-      container.innerHTML = '';
+    container.innerHTML = '';
+    if (newData.status === 404) {
       return;
     }
     const dataJSON = await newData.json();
-
     if (dataJSON.length > 0) {
       for (let elem of dataJSON) {
         container.innerHTML += Comentario({ ...elem });
